+++ client_test.go
@@ -22,7 +22,7 @@ func (stubSigner) Sign([]byte) (*signature.Signature, error) { return &signature
 func callProvision(rawURL string) (string, bool, error) {
 	ctx := context.Background()
 	sURL, _ := url.Parse(rawURL)
-	c := New(sURL, &url.URL{}, stubSigner{})
+	c := New(sURL, &url.URL{}, stubSigner{}, nil)
 
 	cbID := manifold.ID{}
 	resID := manifold.ID{}
@@ -32,7 +32,7 @@ func callProvision(rawURL string) (string, bool, error) {
 func callProvisionCredentials(rawURL string) (map[string]string, string, bool, error) {
 	ctx := context.Background()
 	sURL, _ := url.Parse(rawURL)
-	c := New(sURL, &url.URL{}, stubSigner{})
+	c := New(sURL, &url.URL{}, stubSigner{}, nil)
 
 	cbID := manifold.ID{}
 	resID := manifold.ID{}
@@ -43,7 +43,7 @@ func callProvisionCredentials(rawURL string) (map[string]string, string, bool, e
 func callChangePlan(rawURL string) (string, bool, error) {
 	ctx := context.Background()
 	sURL, _ := url.Parse(rawURL)
-	c := New(sURL, &url.URL{}, stubSigner{})
+	c := New(sURL, &url.URL{}, stubSigner{}, nil)
 
 	cbID := manifold.ID{}
 	resID := manifold.ID{}
@@ -55,7 +55,7 @@ func callDeprovisionCredentials(rawURL string) (string, bool, error) {
 	ctx := context.Background()
 	sURL, _ := url.Parse(rawURL)
 
-	c := New(sURL, &url.URL{}, stubSigner{})
+	c := New(sURL, &url.URL{}, stubSigner{}, nil)
 
 	cbID := manifold.ID{}
 	credID := manifold.ID{}
@@ -67,7 +67,7 @@ func callDeprovisionResource(rawURL string) (string, bool, error) {
 	ctx := context.Background()
 	sURL, _ := url.Parse(rawURL)
 
-	c := New(sURL, &url.URL{}, stubSigner{})
+	c := New(sURL, &url.URL{}, stubSigner{}, nil)
 
 	cbID := manifold.ID{}
 	resID := manifold.ID{}
@@ -105,7 +105,7 @@ func testCreateSSOURL(base, code string) (string, manifold.ID, error) {
 		return "", ID, err
 	}
 
-	c := New(u, &url.URL{}, stubSigner{})
+	c := New(u, &url.URL{}, stubSigner{}, nil)
 	return c.CreateSsoURL(code, ID).String(), ID, nil
 }
 
@@ -228,8 +228,9 @@ func TestProvisionResource(t *testing.T) {
 	})
 
 	t.Run("400 bad request valid response", withCode(http.StatusBadRequest, func(url string) {
-		_, _, err := callProvision(url)
+		msg, _, err := callProvision(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.BadRequestError, "i dont get ya")))
 	}))
 
@@ -242,8 +243,9 @@ func TestProvisionResource(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callProvision(srv.URL)
+		msg, _, err := callProvision(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(`no consumer: "text/html"`))
 	})
 
@@ -256,28 +258,37 @@ func TestProvisionResource(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callProvision(srv.URL)
+		msg, _, err := callProvision(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("401 unauthorized valid response", withCode(http.StatusUnauthorized, func(url string) {
-		_, _, err := callProvision(url)
+		msg, _, err := callProvision(url)
+
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.UnauthorizedError, "i dont get ya")))
 	}))
 
 	t.Run("409 conflict valid response", withCode(http.StatusConflict, func(url string) {
-		_, _, err := callProvision(url)
+		msg, _, err := callProvision(url)
+
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.ConflictError, "i dont get ya")))
 	}))
 
 	t.Run("500 internal server error valid response", withCode(http.StatusInternalServerError, func(url string) {
-		_, _, err := callProvision(url)
+		msg, _, err := callProvision(url)
+
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.InternalServerError, "i dont get ya")))
 	}))
 
 	t.Run("503 service unavailable, unrecognized status code", withCode(http.StatusServiceUnavailable, func(url string) {
-		_, _, err := callProvision(url)
+		msg, _, err := callProvision(url)
+
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(IsFatal(err)).To(gm.BeFalse())
 	}))
 
@@ -291,8 +302,9 @@ func TestProvisionResource(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callProvision(srv.URL)
+		msg, _, err := callProvision(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(IsFatal(err)).To(gm.BeFalse())
 	})
 }
@@ -372,14 +384,16 @@ func TestProvisionCredentials(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, _, err := callProvisionCredentials(srv.URL)
+		_, msg, _, err := callProvisionCredentials(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("400 bad request valid response", withCode(http.StatusBadRequest, func(url string) {
-		_, _, _, err := callProvisionCredentials(url)
+		_, msg, _, err := callProvisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.BadRequestError, "i dont get ya")))
 	}))
 
@@ -392,8 +406,9 @@ func TestProvisionCredentials(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, _, err := callProvisionCredentials(srv.URL)
+		_, msg, _, err := callProvisionCredentials(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(`no consumer: "text/html"`))
 	})
 
@@ -406,26 +421,30 @@ func TestProvisionCredentials(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, _, err := callProvisionCredentials(srv.URL)
+		_, msg, _, err := callProvisionCredentials(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("404 not found valid response", withCode(http.StatusNotFound, func(url string) {
-		_, _, _, err := callProvisionCredentials(url)
+		_, msg, _, err := callProvisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.NotFoundError, "i dont get ya")))
 	}))
 
 	t.Run("409 conflict valid response", withCode(http.StatusConflict, func(url string) {
-		_, _, _, err := callProvisionCredentials(url)
+		_, msg, _, err := callProvisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.ConflictError, "i dont get ya")))
 	}))
 
 	t.Run("500 internal server error valid response", withCode(http.StatusInternalServerError, func(url string) {
-		_, _, _, err := callProvisionCredentials(url)
+		_, msg, _, err := callProvisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.InternalServerError, "i dont get ya")))
 	}))
 }
@@ -503,8 +522,9 @@ func TestChangePlan(t *testing.T) {
 	})
 
 	t.Run("400 bad request valid response", withCode(http.StatusBadRequest, func(url string) {
-		_, _, err := callChangePlan(url)
+		msg, _, err := callChangePlan(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.BadRequestError, "i dont get ya")))
 	}))
 
@@ -517,7 +537,9 @@ func TestChangePlan(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callChangePlan(srv.URL)
+		msg, _, err := callChangePlan(srv.URL)
+
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(`no consumer: "text/html"`))
 	})
 
@@ -530,26 +552,30 @@ func TestChangePlan(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callChangePlan(srv.URL)
+		msg, _, err := callChangePlan(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("401 bad unauthorized valid response", withCode(http.StatusUnauthorized, func(url string) {
-		_, _, err := callChangePlan(url)
+		msg, _, err := callChangePlan(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.UnauthorizedError, "i dont get ya")))
 	}))
 
 	t.Run("404 not found valid response", withCode(http.StatusNotFound, func(url string) {
-		_, _, err := callChangePlan(url)
+		msg, _, err := callChangePlan(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.NotFoundError, "i dont get ya")))
 	}))
 
 	t.Run("500 internal server error valid response", withCode(http.StatusInternalServerError, func(url string) {
-		_, _, err := callChangePlan(url)
+		msg, _, err := callChangePlan(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.InternalServerError, "i dont get ya")))
 	}))
 }
@@ -590,8 +616,9 @@ func TestDeprovisionCredentials(t *testing.T) {
 	})
 
 	t.Run("400 bad request error valid response", withCode(http.StatusBadRequest, func(url string) {
-		_, _, err := callDeprovisionCredentials(url)
+		msg, _, err := callDeprovisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.BadRequestError, "i dont get ya")))
 	}))
 
@@ -604,8 +631,9 @@ func TestDeprovisionCredentials(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callDeprovisionCredentials(srv.URL)
+		msg, _, err := callDeprovisionCredentials(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(`no consumer: "text/html"`))
 	})
 
@@ -618,26 +646,30 @@ func TestDeprovisionCredentials(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callDeprovisionCredentials(srv.URL)
+		msg, _, err := callDeprovisionCredentials(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("401 unauthorized valid response", withCode(http.StatusUnauthorized, func(url string) {
-		_, _, err := callDeprovisionCredentials(url)
+		msg, _, err := callDeprovisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.UnauthorizedError, "i dont get ya")))
 	}))
 
 	t.Run("404 not found valid response", withCode(http.StatusNotFound, func(url string) {
-		_, _, err := callDeprovisionCredentials(url)
+		msg, _, err := callDeprovisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.NotFoundError, "i dont get ya")))
 	}))
 
 	t.Run("500 internal server error valid response", withCode(http.StatusInternalServerError, func(url string) {
-		_, _, err := callDeprovisionCredentials(url)
+		msg, _, err := callDeprovisionCredentials(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.InternalServerError, "i dont get ya")))
 	}))
 }
@@ -678,8 +710,9 @@ func TestDeprovisionResource(t *testing.T) {
 	})
 
 	t.Run("400 bad request valid response", withCode(http.StatusBadRequest, func(url string) {
-		_, _, err := callDeprovisionResource(url)
+		msg, _, err := callDeprovisionResource(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.BadRequestError, "i dont get ya")))
 	}))
 
@@ -692,8 +725,9 @@ func TestDeprovisionResource(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callDeprovisionResource(srv.URL)
+		msg, _, err := callDeprovisionResource(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(`no consumer: "text/html"`))
 	})
 
@@ -706,26 +740,30 @@ func TestDeprovisionResource(t *testing.T) {
 		}))
 		defer srv.Close()
 
-		_, _, err := callDeprovisionResource(srv.URL)
+		msg, _, err := callDeprovisionResource(srv.URL)
 
+		gm.Expect(msg).To(gm.Equal(""))
 		gm.Expect(err).To(gm.MatchError(ErrMissingMsg))
 	})
 
 	t.Run("400 unauthorized valid response", withCode(http.StatusUnauthorized, func(url string) {
-		_, _, err := callDeprovisionResource(url)
+		msg, _, err := callDeprovisionResource(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.UnauthorizedError, "i dont get ya")))
 	}))
 
 	t.Run("404 not found valid response", withCode(http.StatusNotFound, func(url string) {
-		_, _, err := callDeprovisionResource(url)
+		msg, _, err := callDeprovisionResource(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.NotFoundError, "i dont get ya")))
 	}))
 
 	t.Run("500 internal server error valid response", withCode(http.StatusInternalServerError, func(url string) {
-		_, _, err := callDeprovisionResource(url)
+		msg, _, err := callDeprovisionResource(url)
 
+		gm.Expect(msg).To(gm.Equal("i dont get ya"))
 		gm.Expect(err).To(gm.MatchError(NewError(errors.InternalServerError, "i dont get ya")))
 	}))
 }